# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FLIT Core API
  version: v1
  description: |
    Contrato REST canónico de `core-api` (Trámites 2.0).
    Spec en runtime (auto-generada desde Minimal APIs): `GET /openapi/v1.json`
    Features documentadas en esta versión: #9549 (Identidad) y #9550 (Compañías).
servers:
  - url: http://localhost:3030
    description: DEV local
  - url: https://dev.flit.com.co
    description: DEV VPS (ajustar según despliegue)
...
paths:
...
  /api/v1/procedures-config/rules/evaluate:
    post:
      tags: [Procedures Config - Rules]
      summary: Evalúa reglas activas o snapshot y persiste bitácora (HU #9694 AC4)
      operationId: EvaluateProcedureRules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, procedureTypeId, capturedFields]
              properties:
                tenantId: { type: string, format: uuid }
                procedureTypeId: { type: string, format: uuid }
                capturedFields:
                  type: object
                  additionalProperties: { type: string, nullable: true }
                configSnapshot: { type: object, nullable: true }
                procedureInstanceId: { type: string, format: uuid, nullable: true }
      responses:
        "200":
          description: Reglas coincidentes, acciones e invocaciones de endpoints
        "403":
          description: Requiere modulo.tramites.ver
  /api/v1/procedures-config/rules/simulate:
    post:
      tags: [Procedures Config - Rules]
      summary: Simula evaluación de reglas sin persistir ni invocar endpoints (HU #9694 AC2)
      operationId: SimulateProcedureRules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, procedureTypeId, capturedFields]
              properties:
                tenantId: { type: string, format: uuid }
                procedureTypeId: { type: string, format: uuid }
                capturedFields:
                  type: object
                  additionalProperties: { type: string, nullable: true }
                configSnapshot: { type: object, nullable: true }
      responses:
        "200":
          description: Reglas coincidentes, acciones y conflictos detectados
          content:
            application/json:
              schema:
                type: object
                properties:
                  matchedRules: { type: array }
                  actions: { type: array }
                  conflicts:
                    type: array
                    items:
                      type: object
                      properties:
                        conflictType: { type: string }
                        target: { type: string }
                        actionTypes:
                          type: array
                          items: { type: string }
                        ruleIds:
                          type: array
                          items: { type: string, format: uuid }
        "403":
          description: Requiere ui.parametrizacion.simular-reglas o modulo.parametrizacion.crud-total
...
  /api/v1/procedures-config/endpoints:
    get:
      tags: [Procedures Config - Endpoint Catalog]
      summary: Listar catálogo de endpoints externos (HU #9693 field_mapping)
      operationId: ListEndpointCatalog
      parameters:
        - name: tenantId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Incluye field_mapping por endpoint
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sessionCookie:
      type: apiKey
      in: cookie
      name: flit_access
security:
  - bearerAuth: []
  - sessionCookie: []